move wallet functions to cmds_util.py for further refactoring  (#12654)
* move functions to cmds_util.py

* undo mypy change

* add hinting to avoid hassle

* apply change from merge
