+++ chia/cmds/cmds_util.py
@@ -1,5 +1,13 @@
+from typing import Any, Awaitable, Callable, Dict, Optional, Tuple
+
+import aiohttp
+
+from chia.rpc.wallet_rpc_client import WalletRpcClient
 from chia.types.blockchain_format.sized_bytes import bytes32
 from chia.types.mempool_submission_status import MempoolSubmissionStatus
+from chia.util.config import load_config
+from chia.util.default_root import DEFAULT_ROOT_PATH
+from chia.util.ints import uint16
 from chia.wallet.transaction_record import TransactionRecord
 
 
@@ -10,3 +18,99 @@ def transaction_submitted_msg(tx: TransactionRecord) -> str:
 
 def transaction_status_msg(fingerprint: int, tx_id: bytes32) -> str:
     return f"Run 'chia wallet get_transaction -f {fingerprint} -tx 0x{tx_id}' to get status"
+
+
+async def execute_with_wallet(
+    wallet_rpc_port: Optional[int],
+    fingerprint: int,
+    extra_params: Dict[str, Any],
+    function: Callable[[Dict[str, object], WalletRpcClient, int], Awaitable[None]],
+) -> None:
+    try:
+        config = load_config(DEFAULT_ROOT_PATH, "config.yaml")
+        self_hostname = config["self_hostname"]
+        if wallet_rpc_port is None:
+            wallet_rpc_port = config["wallet"]["rpc_port"]
+        wallet_client = await WalletRpcClient.create(self_hostname, uint16(wallet_rpc_port), DEFAULT_ROOT_PATH, config)
+        wallet_client_f = await get_wallet(wallet_client, fingerprint=fingerprint)
+        if wallet_client_f is None:
+            wallet_client.close()
+            await wallet_client.await_closed()
+            return None
+        wallet_client, fingerprint = wallet_client_f
+        await function(extra_params, wallet_client, fingerprint)
+    except KeyboardInterrupt:
+        pass
+    except Exception as e:
+        if isinstance(e, aiohttp.ClientConnectorError):
+            print(
+                f"Connection error. Check if the wallet is running at {wallet_rpc_port}. "
+                "You can run the wallet via:\n\tchia start wallet"
+            )
+        else:
+            print(f"Exception from 'wallet' {e}")
+    wallet_client.close()
+    await wallet_client.await_closed()
+
+
+async def get_wallet(
+    wallet_client: WalletRpcClient, fingerprint: Optional[int] = None
+) -> Optional[Tuple[WalletRpcClient, int]]:
+    if fingerprint is not None:
+        fingerprints = [fingerprint]
+    else:
+        fingerprints = await wallet_client.get_public_keys()
+    if len(fingerprints) == 0:
+        print("No keys loaded. Run 'chia keys generate' or import a key")
+        return None
+    if len(fingerprints) == 1:
+        fingerprint = fingerprints[0]
+    if fingerprint is not None:
+        log_in_response = await wallet_client.log_in(fingerprint)
+    else:
+        logged_in_fingerprint: Optional[int] = await wallet_client.get_logged_in_fingerprint()
+        spacing: str = "  " if logged_in_fingerprint is not None else ""
+        current_sync_status: str = ""
+        if logged_in_fingerprint is not None:
+            if await wallet_client.get_synced():
+                current_sync_status = "Synced"
+            elif await wallet_client.get_sync_status():
+                current_sync_status = "Syncing"
+            else:
+                current_sync_status = "Not Synced"
+        print("Wallet keys:")
+        for i, fp in enumerate(fingerprints):
+            row: str = f"{i + 1}) "
+            row += "* " if fp == logged_in_fingerprint else spacing
+            row += f"{fp}"
+            if fp == logged_in_fingerprint and len(current_sync_status) > 0:
+                row += f" ({current_sync_status})"
+            print(row)
+        val = None
+        prompt: str = (
+            f"Choose a wallet key [1-{len(fingerprints)}] ('q' to quit, or Enter to use {logged_in_fingerprint}): "
+        )
+        while val is None:
+            val = input(prompt)
+            if val == "q":
+                return None
+            elif val == "" and logged_in_fingerprint is not None:
+                fingerprint = logged_in_fingerprint
+                break
+            elif not val.isdigit():
+                val = None
+            else:
+                index = int(val) - 1
+                if index < 0 or index >= len(fingerprints):
+                    print("Invalid value")
+                    val = None
+                    continue
+                else:
+                    fingerprint = fingerprints[index]
+        assert fingerprint is not None
+        log_in_response = await wallet_client.log_in(fingerprint)
+
+    if log_in_response["success"] is False:
+        print(f"Login failed: {log_in_response}")
+        return None
+    return wallet_client, fingerprint
