+++ chia/cmds/wallet.py
@@ -28,7 +28,8 @@ def wallet_cmd() -> None:
 def get_transaction_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: int, tx_id: str, verbose: int) -> None:
     extra_params = {"id": id, "tx_id": tx_id, "verbose": verbose}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_transaction
+    from .wallet_funcs import get_transaction
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_transaction))
 
@@ -108,7 +109,8 @@ def get_transactions_cmd(
     }
 
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_transactions
+    from .wallet_funcs import get_transactions
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_transactions))
 
@@ -175,7 +177,8 @@ def send_cmd(
         "min_coin_amount": min_coin_amount,
     }
     import asyncio
-    from .wallet_funcs import execute_with_wallet, send
+    from .wallet_funcs import send
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, send))
 
@@ -198,7 +201,8 @@ def send_cmd(
 )
 def show_cmd(wallet_rpc_port: Optional[int], fingerprint: int, wallet_type: Optional[str]) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, print_balances
+    from .wallet_funcs import print_balances
+    from chia.cmds.cmds_util import execute_with_wallet
 
     args: Dict[str, Any] = {}
     if wallet_type is not None:
@@ -229,7 +233,8 @@ def show_cmd(wallet_rpc_port: Optional[int], fingerprint: int, wallet_type: Opti
 def get_address_cmd(wallet_rpc_port: Optional[int], id, fingerprint: int, new_address: bool) -> None:
     extra_params = {"id": id, "new_address": new_address}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_address
+    from .wallet_funcs import get_address
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_address))
 
@@ -249,7 +254,8 @@ def get_address_cmd(wallet_rpc_port: Optional[int], id, fingerprint: int, new_ad
 def delete_unconfirmed_transactions_cmd(wallet_rpc_port: Optional[int], id, fingerprint: int) -> None:
     extra_params = {"id": id}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, delete_unconfirmed_transactions
+    from .wallet_funcs import delete_unconfirmed_transactions
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, delete_unconfirmed_transactions))
 
@@ -266,7 +272,8 @@ def delete_unconfirmed_transactions_cmd(wallet_rpc_port: Optional[int], id, fing
 def get_derivation_index_cmd(wallet_rpc_port: Optional[int], fingerprint: int) -> None:
     extra_params: Dict[str, Any] = {}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_derivation_index
+    from .wallet_funcs import get_derivation_index
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_derivation_index))
 
@@ -288,7 +295,8 @@ def get_derivation_index_cmd(wallet_rpc_port: Optional[int], fingerprint: int) -
 def update_derivation_index_cmd(wallet_rpc_port: Optional[int], fingerprint: int, index: int) -> None:
     extra_params = {"index": index}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, update_derivation_index
+    from .wallet_funcs import update_derivation_index
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, update_derivation_index))
 
@@ -322,7 +330,8 @@ def update_derivation_index_cmd(wallet_rpc_port: Optional[int], fingerprint: int
 def add_token_cmd(wallet_rpc_port: Optional[int], asset_id: str, token_name: str, fingerprint: int) -> None:
     extra_params = {"asset_id": asset_id, "token_name": token_name}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, add_token
+    from .wallet_funcs import add_token
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, add_token))
 
@@ -357,7 +366,8 @@ def make_offer_cmd(
 ) -> None:
     extra_params = {"offers": offer, "requests": request, "filepath": filepath, "fee": fee}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, make_offer
+    from .wallet_funcs import make_offer
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, make_offer))
 
@@ -403,7 +413,8 @@ def get_offers_cmd(
         "reverse": reverse,
     }
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_offers
+    from .wallet_funcs import get_offers
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_offers))
 
@@ -425,7 +436,8 @@ def take_offer_cmd(
 ) -> None:
     extra_params = {"file": path_or_hex, "examine_only": examine_only, "fee": fee}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, take_offer
+    from .wallet_funcs import take_offer
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, take_offer))
 
@@ -445,7 +457,8 @@ def take_offer_cmd(
 def cancel_offer_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: str, insecure: bool, fee: str) -> None:
     extra_params = {"id": id, "insecure": insecure, "fee": fee}
     import asyncio
-    from .wallet_funcs import execute_with_wallet, cancel_offer
+    from .wallet_funcs import cancel_offer
+    from chia.cmds.cmds_util import execute_with_wallet
 
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, cancel_offer))
 
@@ -486,7 +499,8 @@ def did_create_wallet_cmd(
     wallet_rpc_port: Optional[int], fingerprint: int, name: Optional[str], amount: Optional[int], fee: Optional[int]
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, create_did_wallet
+    from .wallet_funcs import create_did_wallet
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {"amount": amount, "fee": fee, "name": name}
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, create_did_wallet))
@@ -505,7 +519,8 @@ def did_create_wallet_cmd(
 @click.option("-n", "--name", help="Set the DID wallet name", type=str, required=True)
 def did_wallet_name_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: int, name: str) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, did_set_wallet_name
+    from .wallet_funcs import did_set_wallet_name
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {"wallet_id": id, "name": name}
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, did_set_wallet_name))
@@ -523,7 +538,8 @@ def did_wallet_name_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: in
 @click.option("-i", "--id", help="Id of the wallet to use", type=int, required=True)
 def did_get_did_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: int) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_did
+    from .wallet_funcs import get_did
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {"did_wallet_id": id}
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, get_did))
@@ -549,7 +565,8 @@ def nft_wallet_create_cmd(
     wallet_rpc_port: Optional[int], fingerprint: int, did_id: Optional[str], name: Optional[str]
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, create_nft_wallet
+    from .wallet_funcs import create_nft_wallet
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params: Dict[str, Any] = {"did_id": did_id, "name": name}
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, create_nft_wallet))
@@ -618,7 +635,8 @@ def nft_mint_cmd(
     royalty_percentage_fraction: int,
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, mint_nft
+    from .wallet_funcs import mint_nft
+    from chia.cmds.cmds_util import execute_with_wallet
 
     if metadata_uris is None:
         metadata_uris_list = []
@@ -691,7 +709,8 @@ def nft_add_uri_cmd(
     fee: str,
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, add_uri_to_nft
+    from .wallet_funcs import add_uri_to_nft
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {
         "wallet_id": id,
@@ -734,7 +753,8 @@ def nft_transfer_cmd(
     fee: str,
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, transfer_nft
+    from .wallet_funcs import transfer_nft
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {
         "wallet_id": id,
@@ -757,7 +777,8 @@ def nft_transfer_cmd(
 @click.option("-i", "--id", help="Id of the NFT wallet to use", type=int, required=True)
 def nft_list_cmd(wallet_rpc_port: Optional[int], fingerprint: int, id: int) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, list_nfts
+    from .wallet_funcs import list_nfts
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {"wallet_id": id}
     asyncio.run(execute_with_wallet(wallet_rpc_port, fingerprint, extra_params, list_nfts))
@@ -793,7 +814,8 @@ def nft_set_did_cmd(
     fee: str,
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, set_nft_did
+    from .wallet_funcs import set_nft_did
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {
         "wallet_id": id,
@@ -820,7 +842,8 @@ def nft_get_info_cmd(
     nft_coin_id: str,
 ) -> None:
     import asyncio
-    from .wallet_funcs import execute_with_wallet, get_nft_info
+    from .wallet_funcs import get_nft_info
+    from chia.cmds.cmds_util import execute_with_wallet
 
     extra_params = {
         "nft_coin_id": nft_coin_id,
